Refactor loop internals into smaller focused modules (#36)
Split mod.rs (702→373 lines) and turn.rs (894→753 lines) into
single-responsibility modules: event.rs (AgentEvent, TurnEndReason),
config.rs (AgentLoopConfig), types.rs (internal shared types), and
overflow.rs (emergency context overflow recovery). Public API unchanged.
